Move potting logic into new module
- Potting balls seems like a pretty important feature, so rather than
  tie up all the logic in a single method, I think it deserves an entire
  module.
- To do this, I've created a pooltool/potting module and moved all of
  the code you wrote into pooltool/potting/simple.py.
- I also created a PottingConfig so that different aiming methods could
  easily be plugged in
